Fix: Service worker loads offline content properly when site is first visited
The issue came from the initial loading of the service worker. Since the service worker is just being installed when entering the site, nothing was cached until a refresh was done. The necessary files were inserted into the cache during the "install" event to handle the loading of pages when an individual first visits. They need to actually visit the other pages first before viewing them offline. This is done to preserve memory.